Updated pipeline with new state machine setup and move ros communication functions to roscommm
