+++ catkin_ws/src/ur10_picking/src/scripts/pipeline.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 
-import copy
 import rospy
+from roscomm import ServiceCaller, TopicReader, TopicWriter
 from std_msgs.msg import String
 from std_msgs.msg import Bool
 from geometry_msgs.msg import Pose, PoseArray
@@ -14,10 +14,14 @@ class State:
     Define statemachine class
     Each state will be initiated using this class.
     The class includes state name, on_event, next_state
+    :param: pipeline_core: pipeline_core class containing all pipeline data needed for states
+    :param: event: event causing change to the state
+    :param: state_status: binary status number to used to update state machine status
     """
 
     def __init__(self):
         print("Initiating state:", str(self))
+        self.state_status = 0b00000000000000
 
     def run(self, pipeline_core):
         assert 0, "run not implemented"
@@ -42,8 +46,12 @@ class Initialise(State):
         :return: integer ID of next state
         """
         print("Initialising the system and prioritising items")
+
+        # TODO: Check arm has iniitiated and is in shelf home
+
+        # TODO: Check prioritised item list has been populated
+
         # TODO: Test node communications
-        # Do item prioritisation program here (NOT FOR DEMO)
 
         state_complete = True
         return self.next_state(state_complete)
@@ -55,9 +63,9 @@ class Initialise(State):
 
         if state_complete:
             print("Initialisation complete")
-            return 2  # Move to Calibration
+            return 0b10010000000000  # Move to Calibration
         else:
-            return 1  # Stay in Initialise
+            return 0b00000000000000  # Stay in Initialise
 
 
 class Calibrate(State):
@@ -92,9 +100,9 @@ class Calibrate(State):
 
         if state_complete:
             print("Calibration complete")
-            return 3  # Move to FindShelf
+            return 0b01011000000000  # Move to FindShelf
         else:
-            return 2  # Stay in Initialise
+            return 0b10010000000000  # Stay in Calibration
 
 
 class FindShelf(State):
@@ -128,11 +136,11 @@ class FindShelf(State):
         pose_msg.incremental = False
 
         # Turn on vacuum
-	pipeline_core.vacuumonoff.call(1)
+        pipeline_core.vacuumonoff.call(1)
 
         pipeline_core.pose_publisher.write_topic(pose_msg)
         rospy.sleep(10.0)
-        
+
         # Shelf E pick
         pose_msg = PoseMessage()
         shelf_centre_pose = Pose()
@@ -148,7 +156,7 @@ class FindShelf(State):
         pose_msg.incremental = False
 
         pipeline_core.pose_publisher.write_topic(pose_msg)
-        rospy.sleep(10.0)	
+        rospy.sleep(10.0)
 
         # Shelf E home
         pose_msg = PoseMessage()
@@ -167,8 +175,8 @@ class FindShelf(State):
         pipeline_core.pose_publisher.write_topic(pose_msg)
         rospy.sleep(10.0)
 
-	# Turn off vacuum
-	pipeline_core.vacuumonoff.call(0)
+        # Turn off vacuum
+        pipeline_core.vacuumonoff.call(0)
 
         state_complete = True
         return self.next_state(state_complete)
@@ -180,9 +188,9 @@ class FindShelf(State):
 
         if state_complete:
             print("Shelf found")
-            return 4  # Move to AssessShelf
+            return 0b01011000000000  # Move to AssessShelf
         else:
-            return 3  # Stay in FindShelf
+            return 0b01011000000000  # Stay in FindShelf
 
 
 class AssessShelf(State):
@@ -201,7 +209,7 @@ class AssessShelf(State):
         # TODO: Assess shelf - use vision node topic to extract the centroid of the item
         # TODO: Print centroid of the item to terminal
 
-        state_complete = True
+        state_complete = "Item found"
         return self.next_state(state_complete)
 
     def on_event(self, event):
@@ -209,98 +217,15 @@ class AssessShelf(State):
 
     def next_state(self, state_complete):
         # Move to next state when there is a confirmed centroid for the item
-        if state_complete:
-            print("Assessment complete")
-            return 5  # End
-        else:
-            return 4  # Stay in AssessShelf
-
-
-class ServiceCaller:
-    """
-    Define service class
-    This class is defined for each ROS service interaction
-    The class can send a request to a service and process a response from a service
-    :attribute servicename: string
-        string name for the service as defined in the service definition
-    :attribute service_class: class
-        service class as defined in the service definition and imported from the relevent package
-    """
-
-    def __init__(self, servicename, service_class):
-        self.servicename = servicename
-        self.service_class = service_class
-        rospy.wait_for_service(servicename)
-        # rospy.loginfo("Connected to service:", servicename)
-
-    def call(self, req):
-        """
-        Sends a service request to the service attributed with this class
-        :param req: request to be sent to the class. The data type for this varies based on the service definition
-        :return: returns the response from the service
-        """
-        rospy.wait_for_service(self.servicename)
-        client = rospy.ServiceProxy(self.servicename, self.service_class)
-        response = client(req)
-        return response
-
-
-class TopicReader:
-    """
-    Define topicreaderclass to read data from published ROS topics
-    This class processes data from a ROS topic and returns it
-    :attribute: topic_name - string name of the topic to which data will be published
-    :attribute: data_class - data type / ROS data class for the published data
-    """
-
-    def __init__(self, topic_name, data_class):
-        self.var = None
-        self.topic_name = topic_name
-        self.data_class = data_class
-        rospy.Subscriber(self.topic_name, self.data_class, self.callback)
-        rospy.sleep(1)
-        rospy.loginfo("Reading topic {}...the following data has been read: {}".format(topic_name, self.var))
-
-    def callback(self, data):
-        """
-        Callback function - this is called automatically from the readtopic() function below
-        Sets the self.var variable to the data from the topic
-        :param data: input autofilled through the rospy.Subscriber function in the readtopic() function below
-        """
-        self.var = data
-        print(data)
-
-    def read_topic(self):
-        """
-        :return: returns the data read from the topic
-        """
-        rospy.Subscriber(self.topic_name, self.data_class, self.callback)
-        rospy.sleep(1)
-
-        return self.var
-
-
-class TopicWriter:
-    """
-    Define topicreader class to read data from published ROS topics
-    This class processes data from a ROS topic and returns it
-    :attribute: readname: name of the topic to subscribe and read from. Default to "none" if publishing only
-    :attribute: writename: name of the topic to publish and write to. Default to "none" if subscribing only
-    :attribute: read_data_class: data
-    """
-
-    def __init__(self, topic_name, data_class):
-        self.var = None
-        self.topic_name = topic_name
-        self.data_class = data_class
-        self.pub = rospy.Publisher(topic_name, self.data_class, queue_size=10)
-
-    def write_topic(self, message):
-        """
-        Write to the topic that has been created
-        :param message: Input data to be published to this topic
-        """
-        self.pub.publish(message)
+        if state_complete == "Item found":
+            print("Assessment complete - item found")
+            return 0b01011110000000  # End
+        if state_complete == "Item not found":
+            print("Assessment complete - item not found")
+            return 0b01011100000000  # Go to work order management to update priority list
+        if state_complete == "Shelf not found":
+            print("Shelf not found")
+            return 0b01010000000000  # Rerun calibration
 
 
 class StateSupervisor:
@@ -310,7 +235,7 @@ class StateSupervisor:
 
     def __init__(self):
 
-        self.status = 0
+        self.status = 0b00000000000000
         self.state_initialise = Initialise()
         self.state_calibrate = Calibrate()
         self.state_find_shelf = FindShelf()
@@ -325,16 +250,16 @@ class StateSupervisor:
         :return: None
         """
 
-        if self.status == 0:
+        if self.status == 0b00000000000000:
             self.status = self.state_initialise.run(pipeline_core)
-        if self.status == 1:
+        if self.status in (0b10010000000000, 0b01010000000000):
             self.status = self.state_calibrate.run(pipeline_core)
-        if self.status == 2:
+        if self.status in (0b10011000000000, 0b10011000000000):
             self.status = self.state_find_shelf.run(pipeline_core)
-        if self.status == 3:
+        if self.status in (0b01011000000000, 0b01111110110000, 0b01011111010000):
             self.status = self.state_assess_shelf.run(pipeline_core)
-        if self.status == 4:
-            print("Demo completed")
+        if self.status == 0b01011110000000:
+            print("End of pipeline code so far")
 
     def report_status(self):
         print("Current state: ", self.status)
@@ -349,6 +274,8 @@ class PipelineCore:
         # Initiate Pipeline Node:
         rospy.init_node("pipeline", anonymous=False)
         self.rate = rospy.Rate(10)
+
+        # TODO: create the priortised item list of dictionaries (from Tj's program)
 
         # Initiate topics and services:
         # UR10 control:
